Keep the domain event handler inside the domain
