+++ sample/Vivarni.Example.API/Program.cs
@@ -3,18 +3,17 @@ using Vivarni.Example.Infrastructure.SQLite;
 
 var builder = WebApplication.CreateBuilder(args);
 
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-builder.Services.AddInfrastructureServices(builder.Configuration);
-builder.Services.AddDomainServices();
-builder.Services.AddControllers();
+builder.Services
+    .AddEndpointsApiExplorer()
+    .AddSwaggerGen()
+    .AddInfrastructureServices(builder.Configuration)
+    .AddDomainServices()
+    .AddControllers();
+
 var app = builder.Build();
 
 app.UseSwagger();
 app.UseSwaggerUI();
-
 app.UseDeveloperExceptionPage();
-
 app.MapControllers();
-
 app.Run();
